Revert "Add ConfigsManager class to use SecretsManager and ParameterStore"
This reverts commit 39adf7e5f52b228684e4d044adc50f243b5b0c7c.
